refactor: ♻️ rename sse module to realtime (#326)
